revamp-categories: Implemented the Web and Mobile Screens of AddEmployee and Profile.
