+++ android/local.properties
@@ -1,4 +1,4 @@
-sdk.dir=C:\\Users\\Sameer Gupta\\AppData\\Local\\Android\\sdk
+sdk.dir=C:\\Users\\UseR\\AppData\\Local\\Android\\sdk
 flutter.sdk=C:\\src\\flutter
 flutter.buildMode=debug
 flutter.versionName=1.0.0
